requirements.txt
einops>=0.7.0
gradio
h5py>=3.10.0
huggingface-hub>=0.22.0
imageio
matplotlib
#numpy>=2.0.0
numpy==1.26.4
opencv-python
pandas
pillow>=10.2.0
#protobuf>=4.25.3
scipy
tables
tabulate
termcolor
timm
tqdm
trimesh
#triton>=2.4.0
#torch>=2.4.0
#torchvision>=0.19.0
#torchaudio>=2.4.0
wandb
#xformers>=0.0.26
